% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{add_inputs}
\alias{add_inputs}
\title{Title}
\usage{
add_inputs(model, ..., replace = FALSE)
}
\arguments{
\item{replace}{}
}
\description{
Title
}
\examples{
model <- retrieve_model('nickcullen31/test-r-api-model32233')
input1 <- create_input('EXAMPLE', 'Example', 'numeric', min_value=0, max_value=10, step_value=1, initial_value=5)
input2 <- create_input('EXAMPLE2', 'Example2', 'character', options=c('a','b'), initial_value='a')
model <- model \%>\% add_inputs(input1, input2, replace=TRUE)
}
