% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{add_validation}
\alias{add_validation}
\title{Submit a model validation to climo.ai}
\usage{
add_validation(model, validation, cohort, internal)
}
\arguments{
\item{internal}{}
}
\description{
Submit a model validation to climo.ai
}
