% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climo-package.R
\docType{package}
\name{climo}
\alias{climo}
\alias{climo-package}
\title{R API for climo.ai}
\description{
This package is the R interface for climo.ai. It allows you to create, share, evaluate, and collaborate on models from the climo.ai platform directly from the R language.
}
\keyword{internal}
