% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{create_input}
\alias{create_input}
\title{Title}
\usage{
create_input(
  variable,
  label = NULL,
  type = c("continuous", "categorical"),
  min = NULL,
  max = NULL,
  step = NULL,
  options = NULL,
  initial = NULL,
  is_time = FALSE
)
}
\arguments{
\item{is_time}{}
}
\description{
Title
}
\examples{
input1 <- create_input('EXAMPLE', 'Example', 'continuous', min=0, max=10, step=1, initial=5)
input2 <- create_input('EXAMPLE2', 'Example2', 'categorical', options=c('a','b'), initial='a')
}
