% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a climo model on new data}
\usage{
evaluate_model(model, newdata)
}
\arguments{
\item{newdata}{}
}
\description{
Evaluate a climo model on new data
}
\examples{
model <- retrieve_model('nickcullen31/mixed-effects-model')
newdata <- climo::example_data
results <- evaluate_model(model, newdata)
}
