% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.climo}
\alias{predict.climo}
\title{Perform prediction / inference on a climo model}
\usage{
\method{predict}{climo}(model, newdata)
}
\arguments{
\item{newdata}{}
}
\description{
Perform prediction / inference on a climo model
}
\examples{
model <- retrieve_model('nickcullen31/mixed-effects-model')
newdata <- data.frame(AGE = 75,
                      CDRSB_bl = 2.5,
                      PTGENDER = 'Male',
                      time = c(0,1,2))
result <- predict(model, newdata)
}
