% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{create_input}
\alias{create_input}
\title{Title}
\usage{
create_input(
  variable,
  label = NULL,
  variable_type = c("numeric", "character"),
  min_value = NULL,
  max_value = NULL,
  step_value = NULL,
  options = NULL,
  initial_value = NULL,
  is_time = FALSE
)
}
\arguments{
\item{is_time}{}
}
\description{
Title
}
\examples{
input1 <- create_input('EXAMPLE', 'Example', 'numeric', min_value=0,
max_value=10, step_value=1, initial_value=5)
input2 <- create_input('EXAMPLE2', 'Example2', 'character',
options=c('a','b'), initial_value='a')
}
