% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.climo}
\alias{predict.climo}
\title{Perform prediction / inference on a climo model}
\usage{
\method{predict}{climo}(model, newdata)
}
\arguments{
\item{newdata}{}
}
\description{
Perform prediction / inference on a climo model
}
